Configuration.Seed() now seeds datasource
FINALLY FINALLY getting back to this after an awkward never ending week!

Anyway, created a Model for the Product object and a Seed method to populate the datasource. It does this by embedding products.json as a resource and reading it internally. Possibly not the best way to do it depending on how often products.json might change or how big it could be. One assumes seeding (or even re-seeding) would not be a frequent occurance and as it stands it doesn't but a great wait on package size.